requirements.txt
Django>=3.2
djangorestframework>=3.13
python>=3.8
psycopg2>=2.9